chore: removing commented unnecessary codes
