+++ src/app/[locale]/(dashboard)/blogs/[blogId]/page.tsx
@@ -1,18 +1,7 @@
 import { fetchBlog } from "@/src/lib/data-service";
 import ReactMarkdown from "react-markdown";
-
 import { notFound } from "next/navigation";
-import { supabase } from "../../../../../lib/supabase";
-
-export interface Blog {
-  id: number;
-  created_at: string;
-  img: string;
-  title_en: string;
-  title_ka: string;
-  body_en: string;
-  body_ka: string;
-}
+import { Blog } from "@/src/types/types";
 
 interface BlogPageProps {
   params: Promise<{
