+++ src/app/[locale]/(dashboard)/blogs/page.tsx
@@ -1,20 +1,11 @@
 import BlogCard from "../../../components/BlogCard";
 import { supabase } from "../../../../lib/supabase";
 import { fetchBlogs } from "@/src/lib/data-service";
+import { Blog } from "@/src/types/types";
 
 export const metadata = {
   title: "Blogs",
 };
-
-export interface Blog {
-  id: number;
-  created_at: string;
-  img: string;
-  title_en: string;
-  title_ka: string;
-  body_en: string;
-  body_ka: string;
-}
 
 interface BlogsProps {
   params: Promise<{
