--- src/app/api/auth/[auth0]/route.ts
@@ -1,3 +1,0 @@
-// import { handleAuth } from "@auth0/nextjs-auth0";
-
-// export const GET = handleAuth();
