+++ src/app/components/CheckoutFormCart.tsx
@@ -1,48 +1,3 @@
-// "use client";
-
-// import React, { useState } from "react";
-// import { createCheckoutSession } from "../actions/stripe";
-
-// interface CheckoutFormProps {
-//   uiMode: "hosted";
-//   locale: string;
-// }
-
-// export default function CheckoutForm({
-//   uiMode,
-//   locale,
-// }: CheckoutFormProps): JSX.Element {
-//   const [loading, setLoading] = useState<boolean>(false);
-
-//   const formAction = async (): Promise<void> => {
-//     setLoading(true);
-
-//     const formData = new FormData();
-//     formData.append("uiMode", uiMode);
-//     formData.append("priceId", "price_1QWjnEDzVRyJ9zCn2ZqfDTWs");
-//     formData.append("locale", locale);
-
-//     const { url } = await createCheckoutSession(formData);
-
-//     if (url) {
-//       window.location.assign(url);
-//     }
-
-//     setLoading(false);
-//   };
-
-//   return (
-//     <>
-//       <button
-//         className="px-8 py-3 bg-green-600 text-white rounded-lg hover:bg-green-700 transition"
-//         onClick={formAction}
-//         disabled={loading}
-//       >
-//         Checkout
-//       </button>
-//     </>
-//   );
-// }
 "use client";
 
 import React, { useState } from "react";
