+++ src/types/types.ts
@@ -9,3 +9,13 @@ export interface Product {
   description: string;
   image: string;
 }
+
+export interface Blog {
+  id: number;
+  created_at: string;
+  img: string;
+  title_en: string;
+  title_ka: string;
+  body_en: string;
+  body_ka: string;
+}
